LPS-56848 Move converter to module
