+++ modules/apps/journal/journal-api/src/com/liferay/journal/util/JournalConverter.java
@@ -12,9 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.journal.util;
-
-import aQute.bnd.annotation.ProviderType;
+package com.liferay.journal.util;
 
 import com.liferay.portal.kernel.xml.Document;
 import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
@@ -26,7 +24,6 @@ import java.util.Locale;
  * @author Marcellus Tavares
  * @author Bruno Basto
  */
-@ProviderType
 public interface JournalConverter {
 
 	public String getContent(DDMStructure ddmStructure, Fields ddmFields)
